# packages/api-client/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "HealthTribe AI API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/auth/me": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "Get Current User Info",
        "operationId": "get_current_user_info_api_v1_auth_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/auth/switch-role": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Switch Role",
        "description": "DEV ONLY endpoint to switch roles for testing.",
        "operationId": "switch_role_api_v1_auth_switch_role_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoleSwitchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/v1/appointments/": {
      "get": {
        "tags": [
          "Appointments"
        ],
        "summary": "Get All",
        "operationId": "get_all_api_v1_appointments__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AppointmentResponse"
                  },
                  "type": "array",
                  "title": "Response Get All Api V1 Appointments  Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "Appointments"
        ],
        "summary": "Create Appointment",
        "operationId": "create_appointment_api_v1_appointments__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppointmentCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppointmentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/appointments/{appointment_id}/complete": {
      "post": {
        "tags": [
          "Appointments"
        ],
        "summary": "Complete Consultation",
        "operationId": "complete_consultation_api_v1_appointments__appointment_id__complete_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "appointment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Appointment Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConsultationCompletion"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppointmentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/appointments/{appointment_id}/consultation": {
      "get": {
        "tags": [
          "Appointments"
        ],
        "summary": "Get Consultation",
        "operationId": "get_consultation_api_v1_appointments__appointment_id__consultation_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "appointment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Appointment Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/abha/link/{patient_id}": {
      "post": {
        "tags": [
          "ABHA"
        ],
        "summary": "Link Abha",
        "operationId": "link_abha_api_v1_abha_link__patient_id__post",
        "parameters": [
          {
            "name": "patient_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Patient Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ABHALinkRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ABHAIdentityResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/abha/identity/{patient_id}": {
      "get": {
        "tags": [
          "ABHA"
        ],
        "summary": "Get Abha Identity",
        "operationId": "get_abha_identity_api_v1_abha_identity__patient_id__get",
        "parameters": [
          {
            "name": "patient_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Patient Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ABHAIdentityResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/abha/import/{patient_id}": {
      "post": {
        "tags": [
          "ABHA"
        ],
        "summary": "Import Records",
        "operationId": "import_records_api_v1_abha_import__patient_id__post",
        "parameters": [
          {
            "name": "patient_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Patient Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImportRecordsRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportSessionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ABHAIdentityResponse": {
        "properties": {
          "abha_number": {
            "type": "string",
            "title": "Abha Number"
          },
          "abha_address": {
            "type": "string",
            "title": "Abha Address"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "patient_id": {
            "type": "string",
            "title": "Patient Id"
          },
          "verification_status": {
            "type": "string",
            "title": "Verification Status"
          },
          "verification_method": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Verification Method"
          },
          "linked_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Linked At"
          },
          "is_primary": {
            "type": "boolean",
            "title": "Is Primary"
          }
        },
        "type": "object",
        "required": [
          "abha_number",
          "abha_address",
          "id",
          "patient_id",
          "verification_status",
          "is_primary"
        ],
        "title": "ABHAIdentityResponse"
      },
      "ABHALinkRequest": {
        "properties": {
          "abha_number": {
            "type": "string",
            "title": "Abha Number"
          },
          "abha_address": {
            "type": "string",
            "title": "Abha Address"
          }
        },
        "type": "object",
        "required": [
          "abha_number",
          "abha_address"
        ],
        "title": "ABHALinkRequest"
      },
...
      "AppointmentCreate": {
        "properties": {
          "doctor_id": {
            "type": "string",
            "format": "uuid",
            "title": "Doctor Id"
          },
          "date": {
            "type": "string",
            "title": "Date"
          },
          "time": {
            "type": "string",
            "title": "Time",
            "default": "10:00 AM"
          },
          "type": {
            "type": "string",
            "title": "Type",
            "default": "In-Person"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "type": "object",
        "required": [
          "doctor_id",
          "date"
        ],
        "title": "AppointmentCreate"
      },
      "AppointmentResponse": {
        "properties": {
          "date": {
            "type": "string",
            "title": "Date"
          },
          "time": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Time"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "doctor_id": {
            "type": "string",
            "format": "uuid",
            "title": "Doctor Id"
          },
          "patient_id": {
            "type": "string",
            "format": "uuid",
            "title": "Patient Id"
          },
          "patient_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Patient Name"
          },
          "meet_link": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Meet Link"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          }
        },
        "type": "object",
        "required": [
          "date",
          "status",
          "type",
          "doctor_id",
          "patient_id",
          "id"
        ],
        "title": "AppointmentResponse"
      },
...
      "ConsultationCompletion": {
        "properties": {
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "medications": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/MedicationCreate"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Medications"
          },
          "lab_orders": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/LabOrderCreate"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lab Orders"
          }
        },
        "type": "object",
        "title": "ConsultationCompletion"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ImportRecordsRequest": {
        "properties": {
          "hospital_id": {
            "type": "string",
            "format": "uuid",
            "title": "Hospital Id"
          },
          "consent_record_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Consent Record Id"
          }
        },
        "type": "object",
        "required": [
          "hospital_id"
        ],
        "title": "ImportRecordsRequest"
      },
      "ImportSessionResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "hospital_id": {
            "type": "string",
            "title": "Hospital Id"
          },
          "date": {
            "type": "string",
            "title": "Date"
          },
          "imported_count": {
            "type": "integer",
            "title": "Imported Count"
          }
        },
        "type": "object",
        "required": [
          "id",
          "hospital_id",
          "date",
          "imported_count"
        ],
        "title": "ImportSessionResponse"
      },
      "LabOrderCreate": {
        "properties": {
          "test_name": {
            "type": "string",
            "title": "Test Name"
          },
          "priority": {
            "type": "string",
            "title": "Priority",
            "default": "ROUTINE"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "type": "object",
        "required": [
          "test_name"
        ],
        "title": "LabOrderCreate"
      },
...
      "MedicationCreate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "dosage": {
            "type": "string",
            "title": "Dosage"
          },
          "frequency": {
            "type": "string",
            "title": "Frequency"
          },
          "duration": {
            "type": "string",
            "title": "Duration"
          },
          "instructions": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Instructions"
          }
        },
        "type": "object",
        "required": [
          "name",
          "dosage",
          "frequency",
          "duration"
        ],
        "title": "MedicationCreate"
      },
...
      "RoleEnum": {
        "type": "string",
        "enum": [
          "PATIENT",
          "DOCTOR",
          "ADMIN"
        ],
        "title": "RoleEnum"
      },
      "RoleSwitchRequest": {
        "properties": {
          "role": {
            "type": "string",
            "title": "Role"
          }
        },
        "type": "object",
        "required": [
          "role"
        ],
        "title": "RoleSwitchRequest"
      },
...
      "UserResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "clerk_user_id": {
            "type": "string",
            "title": "Clerk User Id"
          },
          "role": {
            "$ref": "#/components/schemas/RoleEnum"
          }
        },
        "type": "object",
        "required": [
          "id",
          "clerk_user_id",
          "role"
        ],
        "title": "UserResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}